refactor: apply clippy suggestions for cleaner code patterns
- Use `abs_diff()` for unsigned integer difference calculations in connectors
- Remove redundant type casts in transparency calculations
- Simplify conditional logic in base64 decoder padding handling
- Use `clamp()` instead of chained `max().min()` for font size bounds
- Replace `push_str(">")` with `push('>')` for single character
- Use `if let` pattern matching instead of `is_some()` + `unwrap()`
- Remove redundant closures in `map

+++ src/generator/gradients.rs
@@ -93,7 +93,7 @@ impl GradientStop {
 
     /// Set transparency (0-100 percent)
     pub fn with_transparency(mut self, percent: u32) -> Self {
-        self.transparency = Some((percent.min(100) * 1000) as u32);
+        self.transparency = Some(percent.min(100) * 1000);
         self
     }
 }
